test: better structured tests for visibility sims
